pyproject.toml
[project]
name = "sam2streamengine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pydantic>=2.11.4",
    "pyzmq>=26.4.0",
    "rich>=14.0.0",
    "sam2>=1.1.0",
    "torch>=2.7.0",
]
